+++ packages/kobber-components/src/button/Button.stories.ts
@@ -64,11 +64,16 @@ export const Button: StoryObj<Args> = {
     state: states[0],
     icon: buttonIconSettings[1],
     link: false,
+    fullWidth: false,
   },
   parameters: {
     layout: "centered",
   },
-  render: args => renderButton(args),
+  render: args =>
+    args.fullWidth
+      ? `fixed container for previewing full width button
+      <div style="padding: 2rem; width: 80vw; border: 1px solid grey;">${renderButton(args)}</div>`
+      : renderButton(args),
 };
 
 export const Buttons: StoryObj<Args> = {
@@ -191,7 +196,7 @@ const renderVariant = (args: Args) => {
 };
 
 const renderButton = (args: Args) => {
-  const { color, variant, level, state, icon, text, link } = args;
+  const { color, variant, level, state, icon, text, link, fullWidth } = args;
 
   return `
 <kobber-button 
@@ -201,6 +206,7 @@ const renderButton = (args: Args) => {
   level="${level}" 
   ${state === "disabled" ? "disabled" : ""} 
   ${icon === "left" ? "iconFirst" : ""} 
+  ${fullWidth ? "fullWidth" : ""} 
   ${link ? "href='#' target='_blank'" : ""}
   aria-label="optional button label">
   ${text ? text : ""}